[perf] improved testing by running only a single rule
Currently the `valid-jsdoc` is broken. In order to develop the rule for
this branch I wish to be able to focus only on that single rule. For
this reason I have added the command line parameter `single` so that we
may specify which test to run instead of all of them. This has been
documented in the README.