[perf] adding `--lint` and `--no-lint` option to gulp
While developing a rule the linter always stops the build task from
happening. For this reason I believe it is handy to have this option to
skip the linting while on development since some rules such as
`no-unused-variables` may show up while we make adjustments. By default
the linting task is always on, we have to manually specify `—no-lint`
in order to activate it.