[lint] adding the rules from this project
We should think about adding the rules from this project to improve
consistency throughout the code since we will be able to enforce them.
For now I’m adding the `object-curly-spacing` and turning it of since
this will bring bring up the inconsistencies with the import statements
and also it should be discussed in a future issue.