[bug] fixed valid-jsdoc rule (closes #100)
Typescript first tells us that we cannot make `visitClassExpression` as
private since the parent class does not have it as private. When
calling visitNode from the base class we will get to a switch statement
which will run the visitClassExpression class which gets us into an
infinite loop. Will explain better on the PR.