[feat] adding runScript and runTest functions to the test helpers
They allow us to check that we are getting the correct number of
expected errors and their messages as well as a nice error if in the
future a test breaks due to changes. We can also configure the rule on
a script to script basis. This idea was taken from the eslint tests.