[doc] updated README file using the data in rules.js
The documentation should be more consistent now that it is being auto
generated. This already got a few mismatches. For instance, in some
blocks it used extra spaces which are now removed. Others used `not
applicable to Typescript` as opposed to `Not applicable to Typescript`.
I have opted to go with `Not`. The only data that is lost using this is
the `low priority` comment and the `// ou no-construct` comment. This
should be simple enough to add it to the rules object so that we may
incorporate it later.