[bugfix] async processes mess the readme file creation
Added callbacks to each function so that we may add the next execution.
The readme file should generated correctly now. During the addition of
other rules I noticed that the readme file would be modified in places
it shouldn’t. Upon inspection this was due to the fact that some rule
files were not yet written when the creation of the readme file was
started.