[docs:rules] updated rules file so that readme may be generated properly
We should not be the rules table directly on the README file. Instead
we should modify `src/readme/rules.ts` then then run `gulp readme` to
properly generate the table. This commit brings the recent changes to
the README file up to date. Another change done to the gulp file is to
deal with the async issues when reading and writing files, this was not
added after the revert which contained this fix.