[fix] ter-indent for typescript > 2.0 and helper methods update
As explained in the comment we no longer have access to the dot token.
We have to find it. Here I’m getting it directly under the assumption
that the code has 3 children.

The helper method has been updated to show the differences on the
errors we expect instead of simply showing both outputs.