[docs] generating rule documentation from rule metadata
Following a similar style as TSLint we can generate each of the doc
files for each rule by using the metadata attached to the rule. We
should slowly attempt to remove the `usage` property in readme rules
file since this information is better suited for the metadata.

The readme file should be lean only informing us the about the
existence and location of the rule.