[feat] RuleTester
The rule tester will replace the helper file in the tests. A few more
modifications need to be done before the release. For instance I’m not
happy with the `IFailure` interface name. This needs to be changed. All
the tests for the ter-max-len rule have been updated to make use of the
RuleTester.