[fix] ter-indent class method indentation
The eslint code worked because every functionExpression is wrapped
around by a Method declaration. Thus a method would be properly
indented in eslint. In typescript we have to explicitly check for them.
I have refactored the way we check parameters so that it checks the
closing parenthesis as well as the opening curly brace.