[fix] array-bracket-spacing handles comments (closes #162)
I see no obvious way of getting a reference to the comment. To handle
this we manually check if there is a comment in the place where there
should be “spaces”. If we find it we return only those spaces that need
to be removed. That is, we think of the comment as an element.