[refactor-tests] reverting changes in favor of using tslint parseConfigFile
Upon close inspection on their source code I found out that each
tslint.json raw contents go through the `parseConfigFile` function. The
changes I had made were a very hacky (and ugly) way of creating a
configuration file. This brings us back to the original code and all
tests seem to pass. There is no rush in creating a release for this
since this is for our internal tests.