[feat] ter-no-irregular-whitespace
TSLint provides the rule and they named it like one in this project.
Keep in mind that the rule differs from the one in ESLint so we will
keep it in the project under the ter prefix. The tests were failing
because it was referring to the TSLint rule instead of the one provided
here.